workspace/sp500/requirements.txt
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
frozendict==2.4.6
idna==3.10
jmespath==1.0.1
multitasking==0.0.11
numpy==2.2.5
platformdirs==4.3.8
protobuf==5.29.4
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
soupsieve==2.7
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
requests==2.32.3
websockets==15.0.1

beautifulsoup4==4.13.4
botocore==1.38.13             
s3transfer==0.12.0         
boto3==1.38.13                
pandas==2.2.3                 
peewee==3.18.1
curl_cffi==0.10.0             
finnhub-python==2.4.23        
kafka-python==2.2.6
python-dotenv==1.1.0
yfinance==0.2.59              
lxml==5.2.1